feat add&remove functions
